pyproject.toml
[project]
name = "inferdyssey"
version = "0.2.0"
description = "Research OS for Apple Silicon inference — learn, experiment, contribute"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "mlx>=0.22",
    "reflex>=0.8",
    "gitpython>=3.1",
    "openai>=1.0",
    "psutil>=5.9",
    "pyyaml>=6.0",
    "tiktoken>=0.7",
    "plotly>=5.0",
    "requests>=2.28",
    "zvec",
    "sentence-transformers>=2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "inferdyssey"]
